Scripts: formatting, Shellcheck and Bashate enhancements

+++ shutdown.sh
@@ -8,20 +8,20 @@ TAG="octo"
 REMOVE_IMAGE=false
 
 while [ $# -gt 0 ]; do
-  case "$1" in
+    case "$1" in
     -remove|--remove-images)
-      REMOVE_IMAGE=true
-      ;;
+        REMOVE_IMAGE=true
+        ;;
     -p|--project)
-      PROJECT_NAME="$2"
-      shift
-      ;;
+        PROJECT_NAME="$2"
+        shift
+        ;;
     *)
-      echo "Usage: $0 [--remove-image] [-p|--project <project_name>]"
-      exit 1
-      ;;
-  esac
-  shift
+        echo "Usage: $0 [--remove-image] [-p|--project <project_name>]"
+        exit 1
+        ;;
+    esac
+    shift
 done
 
 # Check if the script is run as root
@@ -34,10 +34,14 @@ fi
 echo "Shutting down Docker services..."
 if command -v docker-compose >/dev/null 2>&1; then
     echo "Stopping services using docker-compose..."
-    docker-compose -p "$PROJECT_NAME" -f docker-compose-onionbalance.yaml --profile "$PROJECT_NAME" down
+    docker-compose -p "$PROJECT_NAME" \
+    -f docker-compose-onionbalance.yaml \
+    --profile "$PROJECT_NAME" down
 elif command -v docker >/dev/null 2>&1 && docker compose version >/dev/null 2>&1; then
     echo "Stopping services using Docker Compose v2..."
-    docker compose -p "$PROJECT_NAME" -f docker-compose-onionbalance.yaml --profile "$PROJECT_NAME" down
+    docker compose -p "$PROJECT_NAME" \
+    -f docker-compose-onionbalance.yaml \
+    --profile "$PROJECT_NAME" down
 else
     echo "docker-compose is not installed."
     exit 1
@@ -45,16 +49,16 @@ fi
 
 # Optionally remove project images
 if [ "$REMOVE_IMAGE" = true ]; then
-  echo "Searching for images with tag: $TAG..."
-  IMAGE_IDS=$(docker images | grep "$TAG" | awk '{print $3}')
+    echo "Searching for images with tag: $TAG..."
+    IMAGE_IDS=$(docker images | grep "$TAG" | awk '{print $3}')
 
-  if [ -z "$IMAGE_IDS" ]; then
-    echo "No images found with tag: $TAG."
-  else
-    echo "Removing images..."
-    echo "$IMAGE_IDS" | xargs -r docker rmi -f
-    echo "Images with tag: $TAG removed."
-  fi
+    if [ -z "$IMAGE_IDS" ]; then
+        echo "No images found with tag: $TAG."
+    else
+        echo "Removing images..."
+        echo "$IMAGE_IDS" | xargs -r docker rmi -f
+        echo "Images with tag: $TAG removed."
+    fi
 fi
 
 # Clean up socket directory
@@ -62,7 +66,7 @@ echo "Cleaning up socket directory..."
 SOCKET_DIRECTORY="shared"
 if [ -d "$SOCKET_DIRECTORY" ]; then
     # Remove the socket directory or its contents
-    rm -rf "$SOCKET_DIRECTORY"/*
+    rm -rf "${SOCKET_DIRECTORY:?}"/*
     echo "Socket directory cleaned."
 else
     echo "Socket directory does not exist. No need to clean."
